==> 04-missing.Rmd <==
# Missing values

```{r}
library(readxl)
library(ggplot2)
library(tibble)
library(tidyverse)
library(magrittr)
```


```{r}
fm_ern_new <- read.csv("./data_train/fm_ern_new.csv")
tidyfm <- fm_ern_new %>% 
    gather(key, value, -Year) %>% 
    mutate(missing = ifelse(value=="N", "yes", "no"))

ggplot(tidyfm, aes(x = Year, y =key , fill = missing)) +
  geom_tile(color = "white") + 
  ggtitle("Missing value in fm_ern_new") +
  scale_fill_viridis_d() + 
  theme_bw()
```

After inspecting all transformed data sets, it is good to see only  fm_ern has missing values and the number of missing data isn't large. Also, the majority of missing values gather in the period of 1960-1966, which is less important in this data set. Though some values are missing in 1960-1966, there are still 5 column containing complete data can be still used to study the median earning of different genders. So, the report don't dispose the missing value in 1960-1966 in this part.

In year 1974, there are two missing values in the "ft_workers_num_male" and "ft_workers_num_male" . From observing the data, the similar year has similar value, thus the report fills these two value by using the mean of the worker number from previous and next year. 


```{r}
miss_y<-tidyfm[(tidyfm$missing=="yes" & tidyfm$key=="ft_workers_num_male"),]$Year[1]

above_y <- fm_ern_new[fm_ern_new$Year == (miss_y+1),c("ft_workers_num_male","ft_workers_num_female")]
below_y <- fm_ern_new[fm_ern_new$Year == (miss_y-1),c("ft_workers_num_male","ft_workers_num_female")]

tmp <- rbind(as.numeric(above_y),as.numeric(below_y))
fm_ern_new[fm_ern_new$Year == miss_y, 
           c("ft_workers_num_male","ft_workers_num_female")]=apply(tmp,2,mean)
write.csv(fm_ern_new,"./data_train/fm_ern_final.csv",row.names=FALSE,col.names=TRUE,sep=',')
```
