==> 07-conclusion.Rmd <==
# Conclusion

From the exploratory analysis of each factor's impact on income and employment, we can conclude the following points. First, the rising female share of non-agricultural employment and income is a global trend. Women are trying hard to gain their fair share of power in the workplace worldwide. From this report, the difference of median income among different races can be found, but the good news is all races have an increasing trend in their income before the outbreak of the COVID-19. This increasing trend for income also shows in different types of households. Higher Education can gain higher income. The age group from 45 to 54 has the highest median income. COVID-19 is mentioned a lot of times in this report and through the analysis, it is clear to see that the pandemic has been influencing people's lives economically.
